api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-ai>=0.0.13
python-dotenv>=1.0.0
python-multipart>=0.0.6
google-genai>=0.2.2
google-generativeai>=0.3.2
httpx>=0.27.2
typing-extensions>=4.8.0
chromadb
sqlalchemy
sentence-transformers
faker